+++ Credixware/Internal/Headers/MenuWrapper.h
@@ -1,0 +1,11 @@
+// Created by Credix - Licence: https://creativecommons.org/licenses/by-nc-sa/4.0/ 
+#ifndef MENUWRAPPER_H
+#define MENUWRAPPER_H
+
+#include "Interfaces.h"
+
+void DrawBox() {
+
+}
+
+#endif